Update commonpackages.R
This is one way to load multiple packages in a session without having to retype the require command over and over. 

When using lapply, the character.only=TRUE is important because this means that "multiple_packages" are character strings and not the name of a package. (When loading packages, character.only=FALSE is the default). The object can then be saved using the save() function. 

After using the lapply function, you can double check and see if the packages are loaded by looking for a checkmark next to the package name in the Packages tab.